fix: 인터페이스 제거
1. 불필요한 인터페이스 제거
